Made charts seperate view files
