+++ BoB/Views/Charts/SubmersionChart.swift
@@ -12,14 +12,6 @@ struct SubmersionChart: View {
     let submersionData: [WaterSubmersionData]
     
     @State private var isTimeSeries: Bool = true
-    
-    // pull out the minimum and maximum temperatures
-    private var minTemperature: Double {
-        submersionData.map{ $0.temperature }.min() ?? -2.0
-    }
-    private var maxTemperature: Double {
-        submersionData.map{ $0.temperature }.max() ?? 150.0
-    }
 
     var body: some View {
         // If there is no data, don't plot anything
@@ -34,63 +26,12 @@ struct SubmersionChart: View {
                 .pickerStyle(.segmented)
                 
                 if isTimeSeries {
-                    timeSeries(submersionData: submersionData, minTemperature: minTemperature, maxTemperature: maxTemperature)
+                    SubmersionTimeSeriesChart(submersionData: submersionData)
                 } else {
                     SubmersionVerticalProfileChart(submersionData: submersionData)
                 }
             }
         }
-    }
-}
-
-// make individual views for each type of submersion plot
-// Time Series of Temperature and Pressure
-struct timeSeries: View {
-    let submersionData: [WaterSubmersionData]
-    let minTemperature: Double
-    let maxTemperature: Double
-    let timeStampManager = TimeStampManager()
-    
-    var body: some View {
-        VStack {
-            Chart(submersionData) { item in
-                LineMark(
-                    x: .value("Date", timeStampManager.ISO8601StringtoDate(item.timestamp) ?? Date()),
-                    y: .value("Depth", -1*(item.depth))
-                )
-                .foregroundStyle(.blue)
-                .interpolationMethod(.linear)
-                
-                PointMark(
-                    x: .value("Date", timeStampManager.ISO8601StringtoDate(item.timestamp) ?? Date()),
-                    y: .value("Depth", -1.0*(item.depth) )
-                )
-                .symbolSize(10)
-                .accessibilityLabel("Depth versus Time")
-                .accessibilityValue("Depth: \(item.depth) meters")
-            }
-            .chartYAxisLabel("Depth [m]")
-            
-            Chart(submersionData) { item in
-                LineMark(
-                    x: .value("Date", timeStampManager.ISO8601StringtoDate(item.timestamp) ?? Date()),
-                    y: .value("Temperature", item.temperature)
-                )
-                .foregroundStyle(.blue)
-                .interpolationMethod(.linear)
-                
-                PointMark(
-                    x: .value("Date", timeStampManager.ISO8601StringtoDate(item.timestamp) ?? Date()),
-                    y: .value("Acceleration", item.temperature)
-                )
-                .symbolSize(10)
-                .accessibilityLabel("Temperature versus Time")
-                .accessibilityValue("Temperature: \(item.temperature) degrees Celsius")
-            }
-            .chartYScale(domain: minTemperature...maxTemperature)
-            .chartYAxisLabel("Temperature [°C]")
-        }
-        .frame(height: 300)
     }
 }
