+++ BoB/Views/Charts/SubmersionVerticalProfileChart.swift
@@ -10,10 +10,22 @@ import Charts
 
 struct SubmersionVerticalProfileChart: View {
     let submersionData: [WaterSubmersionData]
-
-    // Zoom state for X and Y axes
-    @State private var xScale: CGFloat = 1.0
-    @State private var yScale: CGFloat = 1.0
+    
+    // pull out the minimum and maximum temperatures
+    private var minTemperature: Double {
+        submersionData.map{ $0.temperature }.min() ?? -2.0
+    }
+    private var maxTemperature: Double {
+        submersionData.map{ $0.temperature }.max() ?? 120.0
+    }
+    
+    // pull out the minimum and maximum temperatures
+    private var minDepth: Double {
+        submersionData.map{ $0.depth }.min() ?? 0.0
+    }
+    private var maxDepth: Double {
+        submersionData.map{ $0.depth }.max() ?? 44.0
+    }
     
     var body: some View {
         Chart(submersionData) { item in
@@ -24,7 +36,6 @@ struct SubmersionVerticalProfileChart: View {
             .foregroundStyle(.blue)
             .interpolationMethod(.linear)
 
-            
             PointMark(
                 x: .value("Temperature", item.temperature),
                 y: .value("Depth", -1*(item.depth))
@@ -33,42 +44,13 @@ struct SubmersionVerticalProfileChart: View {
             .accessibilityLabel("Temperature versus Depth")
             .accessibilityValue("Temperature: \(item.temperature) degrees Celsius")
         }
-        .chartXScale(domain: xDomain())
-        .chartYScale(domain: yDomain())
+        .chartXScale(domain: minTemperature...maxTemperature)
+        .chartYScale(domain: -1*maxDepth...minDepth)
         .chartXAxisLabel("Temperature [°C]")
         .chartYAxisLabel("Depth [m]")
         .frame(height: 300)
         .contentShape(Rectangle())
         .clipped()
-        .gesture(
-            MagnificationGesture()
-                .onChanged { value in
-                    let clamped = max(1.0, min(value, 10.0))
-                    xScale = clamped
-                    yScale = clamped
-                }
-        )
-    }
-    
-    func xDomain() -> ClosedRange<Double> {
-        let temps = submersionData.map { $0.temperature }
-        guard let minT = temps.min(), let maxT = temps.max() else {
-            return 0...1
-        }
-        let center = (minT + maxT) / 2
-        let range = (maxT - minT) / xScale
-        return (center-range / 2)...(center+range / 2)
-        
-    }
-    
-    func yDomain() -> ClosedRange<Double> {
-        let depths = submersionData.map { -1.0 * $0.depth }
-        guard let minD = depths.min(), let maxD = depths.max() else {
-            return 0...1
-        }
-        let center = (minD + maxD) / 2
-        let range = (maxD - minD) / yScale
-        return (center-range / 2)...(center+range / 2)
     }
 }
